derrick/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.7
beautifulsoup4==4.9.1
billiard==3.3.0.23
bs4==0.0.1
celery==3.1.18
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-background-tasks==1.2.5
django-compat==1.0.15
django-crontab==0.7.1
djangorestframework==3.11.0
idna==2.9
kombu==3.0.37
psycopg2-binary==2.8.5
pytz==2020.1
redis==2.10.3
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
